serverstack: relocate embedder resolution
resolveEmbedder / embeddingChunkOptions / the embedder-request type move
into internal/serverstack (ResolveEmbedder / EmbeddingChunkOptions /
EmbedderRequest with exported fields). cmd keeps a type alias + thin
wrappers; the three construction bodies just switch to the exported
field names. No behaviour change.
